refactor(frontend):move object-utils to utils folder
