+++ libs/ui-components/src/model/components/model-viewer/model-viewer.component.ts
@@ -14,7 +14,7 @@ import { OrbitControls } from 'three/examples/jsm/controls/OrbitControls.js';
 import { MatCardModule } from '@angular/material/card';
 import { MatIconModule } from '@angular/material/icon';
 import { MatSnackBar } from '@angular/material/snack-bar';
-import { ObjectUtils } from './utils';
+import { ObjectUtils } from './object-utils';
 
 @Component({
   selector: 'app-model-viewer',
